requirements.txt
snowflake-connector-python==3.11.1
Flask==3.0.3
pymongo==4.8.0
redis==5.0.7
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
dash==2.17.1
statsmodels==0.14.2
scikit-learn==1.5.1
requests==2.32.3
python-dateutil==2.9.0.post0
flask-cors==4.0.1
python-docx==1.1.2